refactor(isEmpty): moved out of extended directory
BREAKING CHANGE: `isEmpty` has been moved out of `extended` folder and under `operator` folder directly
