--- src/add/operator/extended/isEmpty.ts
@@ -1,7 +1,0 @@
-import {Observable} from '../../../Observable';
-import {isEmpty} from '../../../operator/extended/isEmpty';
-import {KitchenSinkOperators} from '../../../Rx.KitchenSink';
-const observableProto = (<KitchenSinkOperators<any>>Observable.prototype);
-observableProto.isEmpty = isEmpty;
-
-export var _void: void;
